ST03-16 POST new declaration function (Heeft ST03-13 nodig)

+++ iOS/Ilmoitus/Ilmoitus/DeclarationLine.h
@@ -12,7 +12,8 @@
 
 @property (strong, nonatomic) NSString *date;
 @property (nonatomic) float cost;
-@property (nonatomic) int subtype;
+// ID of the subtype for this declarationline
+@property (nonatomic) int64_t subtype;
 
 - (instancetype)initWithDate:(NSString *)date SubType:(NSString *)subtype Cost:(float)cost;
